Add ifdef for Hidden Abilities from Pokémon Expansion and change gBattlerAbility back to a u8.

+++ include/battle.h
@@ -815,7 +815,7 @@ extern u16 gMoveToLearn;
 extern u8 gBattleMonForms[MAX_BATTLERS_COUNT];
 extern u32 gFieldStatuses;
 extern struct FieldTimer gFieldTimers;
-extern u16 gBattlerAbility;
+extern u8 gBattlerAbility;
 extern u16 gPartnerSpriteId;
 
 extern void (*gPreBattleCallback1)(void);
